requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
django-crispy-forms>=2.0
python-dateutil>=2.8.2
